Improved the Modified Heap so that it may take any Java Comparable, rather than requiring classes to implement the now removed "OrderedItem" class.
The toString of the ModifiedHeap now outputs its array of items, rather than the array of the orders of its items.

The modified heap now includes a boolean variable determining the "order" of the heap, ie whether smaller objects are closer to the root or larger objects are closer to the root.